ccast-gui/api/requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
PyJWT==2.1.0
python-dotenv==0.19.2
Werkzeug==2.0.1